converted state ints to doubles

+++ src/Agent.h
@@ -8,17 +8,17 @@
 
 class Agent {
 private:
-    std::vector<int> state; // The internal state of the agent
+    std::vector<double> state; // The internal state of the agent
     std::shared_ptr<Node> startingNode;
     std::shared_ptr<Node> primaryGoal;
     std::vector<std::shared_ptr<Node>> secondaryGoals;
     
 public:
     // Constructor
-    Agent(std::vector<int> state, std::shared_ptr<Node> startingNode, std::shared_ptr<Node> primaryGoal, std::vector<std::shared_ptr<Node>> secondaryGoals);
+    Agent(std::vector<double> state, std::shared_ptr<Node> startingNode, std::shared_ptr<Node> primaryGoal, std::vector<std::shared_ptr<Node>> secondaryGoals);
 
     // Getters
-    std::vector<int> getState();
+    std::vector<double> getState();
     std::shared_ptr<Node> getStartingNode();
     std::shared_ptr<Node> getPrimaryGoal();
     std::vector<std::shared_ptr<Node>> getSecondaryGoals();
